+++ openeo/extra/job_management/_manager.py
@@ -157,14 +157,21 @@ class MultiBackendJobManager:
             - get_error_log_path
             - get_job_metadata_path
 
+    :param download_results:
+        Whether to download job results automatically once the job is completed.
+
     :param cancel_running_job_after:
         Optional temporal limit (in seconds) after which running jobs should be canceled
         by the job manager.
+
 
     .. versionadded:: 0.14.0
 
     .. versionchanged:: 0.32.0
         Added ``cancel_running_job_after`` parameter.
+
+    .. versionchanged:: 0.47.0
+        Added ``download_results`` parameter.
     """
 
     # Expected columns in the job DB dataframes.
@@ -193,6 +200,7 @@ class MultiBackendJobManager:
         poll_sleep: int = 60,
         root_dir: Optional[Union[str, Path]] = ".",
         *,
+        download_results: bool = True,
         cancel_running_job_after: Optional[int] = None,
     ):
         """Create a MultiBackendJobManager."""
@@ -203,6 +211,8 @@ class MultiBackendJobManager:
 
         # An explicit None or "" should also default to "."
         self._root_dir = Path(root_dir or ".")
+
+        self._download_results = download_results
 
         self._cancel_running_job_after = (
             datetime.timedelta(seconds=cancel_running_job_after) if cancel_running_job_after is not None else None
@@ -712,16 +722,16 @@ class MultiBackendJobManager:
         :param row: DataFrame row containing the job's metadata.
         """
         # TODO: param `row` is never accessed in this method. Remove it? Is this intended for future use?
+        if self._download_results:
+            job_metadata = job.describe()
+            job_dir = self.get_job_dir(job.job_id)
+            metadata_path = self.get_job_metadata_path(job.job_id)
 
-        job_metadata = job.describe()
-        job_dir = self.get_job_dir(job.job_id)
-        metadata_path = self.get_job_metadata_path(job.job_id)
+            self.ensure_job_dir_exists(job.job_id)
+            job.get_results().download_files(target=job_dir)
 
-        self.ensure_job_dir_exists(job.job_id)
-        job.get_results().download_files(target=job_dir)
-
-        with metadata_path.open("w", encoding="utf-8") as f:
-            json.dump(job_metadata, f, ensure_ascii=False)
+            with metadata_path.open("w", encoding="utf-8") as f:
+                json.dump(job_metadata, f, ensure_ascii=False)
 
     def on_job_error(self, job: BatchJob, row):
         """